% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_met.R
\name{comp_met}
\alias{comp_met}
\title{La fonction comp_met  crée un data frame qui permet de comparer deux métiers à
l'aide de la part des projets de recrutement qu'ils occupent dans chaque région.}
\usage{
comp_met(metier1, metier2)
}
\arguments{
\item{metier1}{nom de métier}

\item{metier2}{nom de métier}
}
\description{
La fonction comp_met  crée un data frame qui permet de comparer deux métiers à
l'aide de la part des projets de recrutement qu'ils occupent dans chaque région.
}
