% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/comp_reg.R
\name{comp_reg}
\alias{comp_reg}
\title{La fonction Comp_reg crée un data frame qui permet de comparer deux régions à
l'aide des métiers qui y font l'objet du plus grand nombre de projets de recrutement.}
\usage{
comp_reg(region1, region2)
}
\arguments{
\item{region1}{nom de région}

\item{region2}{nom de région}
}
\value{
objet de classe comp_region
}
\description{
La fonction Comp_reg crée un data frame qui permet de comparer deux régions à
l'aide des métiers qui y font l'objet du plus grand nombre de projets de recrutement.
}
\examples{

}
