% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df_metiers.R
\name{df_metiers}
\alias{df_metiers}
\title{Description :la fonction df_metiers est utlisée pour la construction de la classe comp_met.}
\usage{
df_metiers(nom1, nom2)
}
\arguments{
\item{nom1}{nom de métier}

\item{nom2}{nom de métier}
}
\value{
Data frame issu de ResMetBE17.La variable "pct" permet de mesurer l'importance des régions (en termes de part des projets de recrutement) pour chacun des deux metiers
}
\description{
Description :la fonction df_metiers est utlisée pour la construction de la classe comp_met.
}
