% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f.R
\name{f}
\alias{f}
\title{Description: la fonction f est utlisée pour la construction de la classe comp_reg.}
\usage{
f(nom1, nom2)
}
\arguments{
\item{nom1}{nom d'une région}

\item{nom2}{nom d'une région}
}
\value{
Data frame issu de ResMetBE17.La variable "pct" permet de mesurer l'importance des métiers dans chacune des deux régions (en termes de part des projets de recrutement)
}
\description{
Description: la fonction f est utlisée pour la construction de la classe comp_reg.
}
