% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g.R
\name{g}
\alias{g}
\title{Title
Description:la fonction f est utlisée pour la construction de la classe comp_reg}
\usage{
g(d, nom1, nom2, nb)
}
\arguments{
\item{d}{}

\item{nom1}{nom de région}

\item{nom2}{nom de région}

\item{nb}{nombre de métiers sélectionnés dans chaque région. Le critère de sélection
est le pourcentage de projets de recrutement.}
}
\value{
Data frame issu de ResMetBE17
}
\description{
Title
Description:la fonction f est utlisée pour la construction de la classe comp_reg
}
