% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.comp_met}
\alias{plot.comp_met}
\title{Title
Description : Méthode plot de la classe comp_met. Le graphique permet de comparer deux métier à
l'aide de la part des projets de recrutement qu'ils occupent dans chaque région}
\usage{
\method{plot}{comp_met}(obj)
}
\arguments{
\item{obj}{objet de classe comp_met}
}
\value{
diagrammes en barres.
}
\description{
Title
Description : Méthode plot de la classe comp_met. Le graphique permet de comparer deux métier à
l'aide de la part des projets de recrutement qu'ils occupent dans chaque région
}
