% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.comp_reg}
\alias{plot.comp_reg}
\title{Title
Description : Méthode plot de la classe comp_reg. Le graphique permet de comparer deux régions à
l'aide des métiers qui y font l'objet du plus grand nombre de projets de recrutement.}
\usage{
\method{plot}{comp_reg}(obj)
}
\arguments{
\item{obj}{objet de classe comp_reg}
}
\value{
diagrammes en barres.
}
\description{
Title
Description : Méthode plot de la classe comp_reg. Le graphique permet de comparer deux régions à
l'aide des métiers qui y font l'objet du plus grand nombre de projets de recrutement.
}
